woring version with couchdb views for querying time periods

+++ dataprovider/delete-analysedArchives.js
@@ -1,10 +1,11 @@
 var cradle = require('cradle');
 var conf = require('config');
 var _ = require('underscore')._;
+var helper = require('./twapper_modules/helper.js');
 
 var conn;
 var db;
-getDBConnectionFromConfig(conf.couchdb, function(error, database, connection){
+helper.getDBConnectionFromConfig(conf.couchdb, function(error, database, connection){
   if(error == null){
     conn = connection;
     db  = database;
@@ -12,10 +13,11 @@ getDBConnectionFromConfig(conf.couchdb, function(error, database, connection){
     db.all(function(err, res){
       
       _.each(res, function(entry){
-        if(entry.id.indexOf("-") != -1){
+        if(entry.id !== "config" && entry.id.indexOf("_design") == -1){
           db.remove(entry.id, entry.value.rev)
+          console.log("removed: ",entry);
         }
-        console.log("conn.all()",entry);
+        
       });
     })
     
@@ -24,37 +26,6 @@ getDBConnectionFromConfig(conf.couchdb, function(error, database, connection){
   }
 });
 
-function getDBConnectionFromConfig(dbConf, callback){
-  var connection = new(cradle.Connection)(dbConf.host,dbConf.port, dbConf.options);
-
-  connection.info(function (err, json) {
-    if (err == null){
-      console.log("Couchdb ("+json.version+") at "+dbConf.host+" connected.");
-      var forenDb = connection.database(dbConf.dbname); 
-      // Create the db if it doesn't exist.
-      forenDb.exists(function (error, exists) {
-              if (!exists) {
-                forenDb.create( 
-                  function(error, res){
-                    if(error == null){
-                      callback(null, forenDb, connection);
-                    }else{
-                      callback(error, null,null);
-                    }
-                  });
-              }else callback(null, forenDb, connection);
-      });
-    }
-    else{
-      err.host = dbConf.host;
-      callback(err, null);
-      //console.log("info() failed to connect to couchdb : ", err);
-    }
-  });
 
 
-}
 
-function exist(param){
- return !(typeof param === 'undefined');
-}
